requirements.txt
fastapi
moralis
python-dotenv
firebase-admin
